Moved method to parent class.
